refactor(18944): refactor the suggestion list component - rename the file/component - redesign the container and items
