+++ hivemq-edge/src/frontend/src/extensions/datahub/components/interpolation/interpolation.utils.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect } from 'vitest'
 import { getItems, parseInterpolations } from '@datahub/components/interpolation/interpolation.utils.ts'
-import { MentionNodeAttrs } from '@datahub/components/interpolation/MentionList.tsx'
+import { MentionNodeAttrs } from '@datahub/components/interpolation/SuggestionList.tsx'
 
 describe('parseInterpolations', () => {
   it('should convert acceptable placeholders to HTML markup', async () => {
